add typescript package to resolve the warning message from typescript-eslint-parser "WARNING: You are currently running a version of TypeScript which is not officially supported by typescript-eslint-parser" (#1002)
detective-typescript in bit-javascript uses typescript-eslint-parser package.
This package doesn't have typescript as a dependency, instead, it has it as a peerDependency. 
It also expects that the installed typescript will be in a specific range. (currently, only 2.7.x. can't be 2.6.x and can't be 2.8.x).

Sadly, having "typescript: 2.7.0" in bit-javascript dependencies doesn't solve the issue, because there is another package "prettier-eslint" used by Bit, which requires typescript as well.

So, from Bit perspective, it looks like this:
- bit-javascript 
--- typescript@2.7.0
- prettier-eslint
--- typescript@2.6.0

Which one of the typescript is going to be at node_modules root? it's unclear.
If the typescript of prettier-eslint has been chosen to be on the root, you'll get this warning.

The workaround done by this commit is to enforce typescript of the specific version 2.7.0 to be on the root of Bit. It's not ideal of course.