refactoring - strip and add the originallySharedDir in one place only (#1220)
remove many places where the sharedDir were added and stripped. 
Instead, move the strip originallySharedDir to be in one place only - when the consumerComponent is created.
Similarly, move the add originallySharedDir to be in one place only - when the consumerComponent is converted to Version instance.
Also, extracted some methods from Scope class into the Consumer class. When a consumer-component is retrieved from the consumer it always has the sharedDir stripped, regardless whether it is loaded from the filesystem or from the model.