import components objects that were installed as packages upon bit import (#3293)
An author might require bit-components that were installed via a package-manager. in that case, the objects are not imported until bit build or bit tag was running.
This PR makes sure to get the objects on bit import (with no ids), so then in the UI, they'll be shown nicely.