enable export to multiple scopes even when components have circular dependencies (#3325)
There are two issues currently when exporting components to different scopes.

1. a bug. when the components are new and have different defaultScope, it changes the dependencies scope to the component scope unexpectedly.
2. by design. when there is a circular dependency between the components from different scopes, it throws an error that the toposort failed and that is unable to export.
This PR fixes both issues.

To enable circulars, it sends the dependencies along with the components. So then scopeA doesn't need to go to scopeB to import the dependencies, it gets them from the client already.
It introduces a new issue. What if after exporting to scopeA, scopeB fails. Then, scopeA has a component that its dependencies don't exist. We'll fix this issue in the next phase of this task.