fix fetch to not collect parents of dependencies (#3398)
when running bit import, it runs importManyWithAllVersions, which goes to scope.fetch() and asks for all components in .bitmap with allDependencies.
during this fetch, it only fetches the parents of the components you asked. For the dependencies, it runs the _getExternalManyWithoutDependencies, which doesn't collect parents.

The problem started once `fetch()` has all components and runs `componentsToComponentsObjects()`, which didn't distinguish between the components and the dependencies. Only the components were supposed to be fetched with parents, not the dependencies. This PR fixes it to not fetch parents for dependencies.