throw a descriptive error when user env is compiled by non-aspect env (#3470)
Only envs that are configured in workspace.jsonc to use core "aspect" env (or any composition of that env) are valid. Currently, the user has no indication when the env is invalid, it falls back to the default "node" env.

Also, fix workspace.get to not catch any error when loading consumerComponent, instead catch only errors indicating that the component doesn't exist. otherwise, the components returned from this method are model-components unexpectedly.