fix IncorrectEnvAspect to not be thrown when bit defaults to the node-env (#3509)
The IncorrectEnvAspect is thrown when a custom env is compiled by a non-aspect core env. However, when the node-modules is empty, all components include custom-env are default to node-env, which is not an aspect env.
This PR prevents the error to be thrown in such a case and let Bit compiles the custom-envs.