remove the flattenedDevDependencies, combine it with flattenedDependencies (#3541)
The following scenario helps understanding why flattenedDevDependencies has been removed.
compA has compB as a dependency and compC as a devDependency. compB has compC as a dependency.
Currently, when we calculate the flattened dependencies on tag, we add compC in the flattenedDependencies of compA because the entire chain from A to C through B are prod dependencies, not dev. 
As a result, we end up with a Version object that has C as a dev-dependency but the flattenedDevDependencies is empty and an error is thrown that the object is invalid.

In practice, we don't really need this differentiation. The need for the flattened dependencies is to be able to cache all dependencies of a component on the remote, and for this, it doesn't matter whether the dependency is dev or not.

Also, improve Version validation errors by adding the component-id to the error message.