remove "multiple" flag from bit update-dependencies (#3628)
We don't want this to be running from the original scope (like bit-sign).
This command tags or snaps the results, a process that takes some time due to the build pipeline. If we start the tag on the original scope, build and then save the tag to the filesystem, we might get another tag during the process and our tag could override it.