stabilize bit to handle component-not-found gracefully (#3656)
### Background
Bit protects users from the event when dependencies from other users/scopes are deleted by saving them in the same scope of the dependent.
In many cases, though, Bit goes to the original scope and throws ComponentNotFound, leaving the workspace broken with not much to do.
This happens not only when the dependency scope/component is gone, but also when that scope was deleted due to issues and re-created with the same components but different versions. 
The dependents have versions that use the old dependencies versions exposing them again to the same errors of ComponentNotFound or VersionNotFound.

### Import Changes
The goal in this PR was to not break unless it's really needed, which is mostly during tag.
When one scope asks for a component from another scope, always go to the direct dependency scope and fetch all its flattened dependencies (which is, direct and indirect dependencies) from that scope. Avoid going to the indirect dependency scopes because the component might not be there anymore  or the version in question might be missing.
This is applicable for any connection between two scopes. Either between the local and the remote (bit import) or between two remotes (happens as part of bit-export where a remote makes sure it has all flattened dependencies).

### Model Changes
Think about the following scenario,  scopeA/compA@0.0.1 -> scopeB/compB@0.0.1, later, compB was deleted and re-created with 0.0.2.
A user import compA and compB. When compB is received from scopeB it has only 0.0.2. When compB is received from scopeA as part of the dependencies of compA, it comes with 0.0.1. 
Currently, these two components were overwriting each other unpredictably. The last one that was written won.
To fix this, the Component model got a new property `orphanedVersions` ({ tag: hash }), which stores the tags that come from non-original scopes. 
Upon import, we compare the remote we got the component from and the scope-name. If they're the same, the component came from the original scope and is considered as the source-of-truth. Otherwise, the component is coming from the cache of a dependent and its versions (if non-exist locally) go to the orphanedVersions object.
This way, we keep references of these tags that don't exist on the original scope, and we're able to load the associated Version object without braking the application.

### ComponentNotFound new behavior
* when a direct dependency is not found in its scope - stop everything. (throw ComponentNotFound). 
* when a non-direct dependency is not found neither in the dependent nor in its scope - stop tagging, allow the rest. bit-import completes successfully and shows a warning message about the missing dependencies per component.
* when a non-direct dependency is missing from the origin but exists on the dependent - bring it from the dependent.

### Some more fixes in this PR:
* `--all-version` flag of `bit export` is now working on Harmony.
* if the remote of the currently exported component doesn't have it and some versions are missing, it throws an error suggesting to run the export with `--all-versions` flag.
* improve the error when a version exists in the component-model but not on the filesystem.
* throw a descriptive error when an artifact file object is missing
* simplify the import/fetch process and remove some recursive code.
* refactoring: extract and organized the logic of ModelComponent merge into a new class `ModelComponentMerger`.