improve import performance by importing only the delta between two scopes (#3727)
Currently, `bit import`/`bit fetch --components` is slow, mostly because we fetch all components with their history even when part of all history exist locally.

This PR introduces a new "fetch" api type:  `component-delta`, which bring only the missing part in the history per component. The way how it is done is as follows:
A local scope has a 'ref' file per remote+lane with all ids and their head when last imported. If the component has head in this file, it is used later as a stop-hash to not traverse beyond this hash. Otherwise, we bring all history.

Other improvements were done in this PR as well.
1. `bit import` used to bring all dependencies of the latest, now it brings them for the used version.
2. `sources.getMany` got called multiple times for the same components. fixed.
3. the conversion from Version to consumer-component happened multiple times for the same components. fixed.
4. the import/fetch APIs are clearer. `fetch` is always local only. `import` can bring from a remote.