clear scope cache upon .bitmap changes during "bit watch" to avoid downgrading versions (#3908)
Currently, when bit watch is running in the background and user tags a component multiple times, at some point, the watches changes the component version in the .bitmap file to a previous version.
This is happening because the watcher loads the model-component from the cache, which exist in an older version. At this point, it assumes that the .bitmap file is out-of-sync and need to be changed to the latest version of the model-component.
This PR clears the scope cache on any .bitmap change, which fixes the issue.