change fetch implementation to use streams to keep the memory usage low (#3916)
The consumer of the fetch API already uses streams, see #3909.
This PR is to make a similar change for the Fetch API implementation, which normally happens on the remote.

Currently, all objects are cached in-memory and pushed all in once into the client. For large scopes this can cause out-of-memory.
This PR changes the implementation to use streams instead. It implements a Readable stream, which processes the objects that need to be sent and pushes them as soon as they're ready.
With this change, the fetch API returns stream.Readable, instead of ObjectList. Here is how it works for each one of the network protocol.
HTTP - the fetch API returns Readable, then, it is piped into tar-stream, which eventually piped into the response.
FS - no need for tar-stream here. The Readable is piped directly to the Writable.
SSH - still uses the old format, which converts the objects into a huge JSON string, so once the Readable is obtained from the fetch api, it converts it to ObjectList and sends it. As such, it doesn't get this memory improvement.