fix multiple fetching during tag (#3967)
Due to a recent change, bit import re-fetch a component when its Version object has a build-status of Pending/Failed.
During tag, the build-status is pending, as such it tried to re-import it again and again from a remote. In this PR it checks whether a version exists only locally and then it won't try to re-fetch it.

Another issue was that getExternalMany() used to go to the remotes for each one of the components for its flattened-dependencies. With this PR it first accumulate all ids and then go to the remote once.