implement bit-eject for Harmony (#3957)
On Harmony the eject operation is much simpler. No need for the complex mechanism of the rollback, mostly because there are no changes to different package.json files.
A new aspect "eject" was created on the workspace scope with the implementation of eject for Harmony.
A new flag was introduced --keep-files to let the user the option to keep the component files in the workspace.