fix unhandled promise reject during fetch process (#3968)
Fetch process sends the objects to a queue to be written to the filesystem and doesn't wait to completion.
Originally it was done to make the process faster and let the queue control the concurrency. However, it causes issues with error handling.
This PR changes the implementation to wait for the queue to complete processing before continuing to the next object.

The performance had been tested before and after this change and there was no noticeable difference. (on bit-bin with more than 10K objects it took 1:10 min in both cases, which is mostly the download time)