fix loading aspects of workspace and scope together (#4103)
currently, when workspace.loadAspects() loads aspects from both workspace and scope, it fails to load the scope aspects from the workspace and it falls back to load them all from the scope. However, once it does that, the workspace component fails.
In this PR, the components are determined before whether they're workspace or scope and are loaded by the workspace/scope accordingly.