clean the dist directory inside node-modules before writing the new dists during compilation process (#4177)
Currently, if component files were deleted, the dist directory inside `node_modules/<component-name>` still has them, which leads to confusion in some scenarios. 
With this fix, during `bit compile` (or `bit watch` as long as it rewrite the entire component), before it writes the dists files, it removes the directory.