fix bit-status when importing scss files (regression due to #4135) (#4178)
As a result of not symlinking all files in the workspace one by one, but symlinking the entire dir to 'src' folder (see #4135 ), bit status showed errors when importing scss files.
The reason is that scss files are imported using the entire file-path inside the component and not from the main index file. This path didn't exist anymore.
This PR takes care of these files and symlink only them. The rest are handled with the src dir symlink.