roll back #4135 for Linux users. symlink _src dir from node_modules to source files for Windows users only (#4182)
Practically, roll back #4135 changes for Linux users and apply them only for Windows users.
Also, change the dir of src to _src, to not conflict with a potential user's src dir.

To sum it up, with this change:
Linux users -> symlink each one of the files. No dir symlink.
Windows users -> symlink each one of the files if possible (admin), otherwise, hard-link. Also, symlink _src to the entire dir.