fix "bit tag -a <version>" to bump by patch auto-tagged and not by the specified version (#4211)
Currently, when running `bit tag -a 1.0.0` and some component are auto-tagged, they get this "1.0.0" version. 
This PR fixes this regression (was working until a few months ago) to bump the auto-tag by a patch.

Another issue this PR fixes is when using `--minor` or `--major` flag, it used to bump the auto-tagged component according to the flag. Now it always bumps with patch. 

See the comments of #4001 for more details. (this PR does the opposite of the original request).