refactoring, rewrite the entire component-issues (#4269)
Currently, `ConsumerComponent.issues` is a simple object, which makes it hard to maintain. We have lots of if/else in many places, e.g. when serialize/deserialize the content, when printing it in bit-status, when deciding whether it should block the tag and so on.

This PR organizes this `component.issues` to be more OO style. It implements a `Template` pattern with a base class `ComponentIssue` and subclasses of the different issues. 
It makes it easier to implement new issues. It just about extending `ComponentIssue` and filling the blanks when needed. 

Also, all these classes were extracted to a new component on the Harmony code `teambit.component/component-issues`. This way it's possible to easily get the issues details via graphql and later show it in the UI.