fix copy-non-supported-files during build to happen in the pre-build phase and only for the seeders (#4304)
and only for the seeders.

Currently, for Babel compiler, which compiles only seeders and not the entire graph, the copy-files mechanism runs on the entire graph. 
In cases where a component has one env and a dependency has another env and one of them configured to copy-non-supported-files and the other does not, the file is always copied. 
This PR fixes it by running the copy process on the seeders only. 
Also, it runs it on the pre-build phase to avoid cases when TS finds other components by their dists and fails with an error about missing scss files. (it basically fixes the issue caused #4300 to fail).