POC - introduce bit prompt to speed up commands and cut the loading bit time to zero (#4422)
This is an experimental feature. 
The idea is to have a command line tool, similar to typing `node` in the cli, you can type `bit cli` and get `bit >` prompt, from there, you can type commands directly without `bit` prefix. 
The big plus is that bit is loaded into the memory once and doesn't need the bootstrap process for every command.
It feels faster especially when using the auto-complete, where users expect to get feedback immediately.