Create a separate pipelines for tag and snap (#4451)
Currently, both use the deploy-pipeline. This pipeline has been deprecated and two new pipelines were created: tagPipeline and snapPipeline.
By default, the publish task is register to the tag-pipeline only. Users can register it to the snap-pipeline in their custom envs.

Also, introduced a new `--list-tasks` flag for `bit build` to show all tasks for each of the pipelines: build, tag and snap.