fix getting old versions of components from the workspace (#4472)
Currently, when asking the Workspace aspect for an older version of a component, we get a mix of the current version with the old version data, which later on, result in ComponentNotFound error. 
This is a regression started since https://github.com/teambit/bit/pull/4414, which we ignored the scope/version when loading a component.
In this PR, a different approach was taken for the out-of-sync issue. Instead of ignoring the scope/version, we save the old-id and the new-id in case of out-of-sync, so later on, we could retrieve the updated data of that id.