bit-new: import the latest component if --aspect component has no version (#4746)
currently, when `bit new --aspect <component-id>` has no version, it load the latest version from the cache. If a new version was published, it doesn't check for it and just reload the last one. 
This PR re-imports the aspect in this case to make sure it always has the latest.