export current lane to its tracked remote-scope (#4802)
Currently, it's required to specify a remote when exporting lanes. 
With this change, Bit uses the tracking data from scope.json to get the remote-scope of the current lane.
Also, the options to export multiple lanes is disabled. Only the current lane is allowed to be exported. 
The `--lanes` flag is removed, as it's not needed anymore.