fix an error "Cannot read property hash of null"  (#4956)
Two different scenarios causing this error, both were fixed:

*  when component dependencies, which are installed as packages and their objects are imported don't have the head object. The reason for the missing the head is probably because they are needed in previous versions and the head version wasn't required. 
The fix is to not run the get-diverge-data on components that are not in the workspace. It's simply not needed.

* when a tag was interrupted (or error was thrown) toward the end of the build pipeline. It happened because the component object was saved to the filesystem with the new tag and new head, however, the Version object of that head was not saved yet. 
The fix is to avoid saving the component objects when fetching objects from remote when these objects are pending to be saved during tag.