bit-build: remove publish/pack dry-run task (#5020)
This dry-run task, runs `npm publish --dry-run` and `npm pack --dry-run`, which are mostly to show the user some output, but not a real test to make sure the publish will succeed. 
It takes long time to complete and the value it provides doesn't justify the performance penalty.

Also, configure the pack task to run in parallel or 10 at a time.