fix(merge-pending): better handling diverge components (#5101)
Currently, when components have diverged from a remote, they're shown in `bit status` as "pending merge" with a suggestion to use `bit merge <id>`.
This `bit merge` command, creates a snap-merge with two parents, the local and the remote. While this is a good solution to keep the local snaps/tags, it is confusing for users to see the hash in the output of `bit merge` and later on in `bit status` and in the .bitmap file.
This PR, suggests running `bit untag` instead. It also fixes `bit untag` to point to the remote head instead of the previous head. This way, after bit-untag is running the component model locally is the same as the remote and the status is `modified`.