fix(create): don't change newly created component scope from existing component in workspace.jsonc (#5108)
This is related to the out-of-sync mechanism described in https://github.com/teambit/bit/discussions/5074. Specifically the last case. If the added component (during `bit add` or `bit create`) doesn't exist in .bitmap, and a similar component exists in the scope, we assume this is the same component and we re-adding the one from the scope to the .bitmap.
In case `bit create --scope` was used, and we know that the entered scope is different than the scope-name of the component found in the model, no need to make changes to the added component and we add it as new.