chore(create): improve "bit create --scope" to save the scope as the defaultScope (#5124)
Currently, upon creating a component, the defaultScope is determined by the workspace.jsonc file. With this PR, the `--scope` flag is taken into account and is used later to export to this scope.

Implementation wise, a new prop is added to the .bitmap file `defaultScope` and gets deleted upon `bit export`.