fix(lanes): avoid replacing the version in .bitmap with the lane version when switching to a remote lane (#5228)
Normally, when snapping on a lane or switching to a non-main lane, the version is written to the workspace-lane, not to the .bitmap. The .bitmap file represents the main lane. always.
This PR fixes the case when switching lanes from a remote lane to not save the snap inside the .bitmap file.