fix(lanes): show the correct staged versions on bit-status when a component and a lane have different scope (#5294)
Currently, if a component belong to scope X and the lane belong to scope Y, Bit tries to find the Ref file of scope X incorrectly. As a result, because the Ref file is not found, Bit assumes that the component was never exported and all its versions are staged. 
This PR fixes it by searching for the correct Ref file according to the tracking-data in scope.jsonc file.