feat: merge lanes during bit-import with no args (#5306)
Currently, when on a lane, `bit import` (with no args) only updates the ref files (`.bit/refs/remote/<lane-name>`). Not the lane objects. To update the lane objects, `bit lane merge` command is needed.
This is not very intuitive because on main `bit import` does merge the objects when possible and to get the files up-to-date in the filesystem, all needs to be done is `bit checkout latest`.
This PR aligns the behavior of lanes and main to be the same.