fix(aspect-loading): avoid overriding aspects that were set by bit-env/bit-aspect-set by workspace variants (#5342)
For example, if an env was set using `bit env set` command to EnvX and the variants has a rule for this component that sets the env to envY, then, after tagging the component, because the `config` in .bitmap is gone, the env is changed back to envY.
This PR fixes it by marking the changes done by `bit env set` or `bit aspect set` with `__specific: true`, so we know not to override them by the variants. 

Also, `bit aspect list` now shows the aspect-level per aspect, whether they're component level or workspace level. With `--debug` it also shows the origin.