fix(checkout): avoid stripping the newline when running Git to merge files (#5355)
Currently, in case of a merge, e.g. when a file has conflicts, then in the written file, the last "\n" is removed.
The culprit was `execa`, which returned the data from Git without the last newline. This PR fixed it by changing the configuration of execa to not strip this newline.