fix: allow specifying id without scope-name when "pattern" is required (#5377)
Needed for `bit env set` for example, when `<pattern>` is required.
Currently, since it uses `multimatch` to match the id, if the scope-name is omitted, it won't match, unless it's prefixed with `**/`. 
This PR checks whether it's a pattern, and if not, it allows resolving the id without the scope-name.