feat: introduce "--editor" flag for "bit tag" to enable editing tag messages per component (#5402)
Currently, it's not possible when tagging a component to specify a message per component.
With the new `--editor` flag, it'll open an editor, default to `vim` with the possibility to enter a message per component, or a default message to catch all empty messages.
It also provides an option to specify the messages for auto-tagged components with a pre-filled default ("bump dependencies versions").

The editor is what the user entered in the `--editor` flag. 
If none was specify, it'll use the `$EDITOR` or `$VISUAL` os variables. 
If they're empty, it'll default to `vim`.