fix: clear dependencies cache before tagging (#5467)
The trigger for this change was a hard-to-debug case, where the first `bit tag -a` failed, and the followed `bit tag -a --skip-auto-tag` threw an error about non-existing dependency version.
The reason was that the first `bit tag -a` saved into the cache the dependencies according to the tag with auto-tag, so the dependents were bumped. Later, the `bit tag -a --skip-auto-tag` used these bumped version from the cache.