fix: eject .bitmap config upon "bit eject-conf" (#5433)
Currently, `bit eject-conf`, only ejects config from the scope, not from the workspace. As a result, changes in .bitmap are not reflected in the component.json. 
This fixes the issue by getting the aspects from the workspace and when the config is written, it ignores the variants. It also deletes the `config` from the .bitmap once it is ejected.