fix: avoid overriding components after merge from main by the remote (#5623)
When a lane is merged into master, there is no indication in the `Component` object that there is a local change, such as the `state.local` created by snap/tag. 
As a result, when running `bit import` afterwards, the `Component` object gets replaced by the remote one and the merge data is gone. 
This fixes it by marking the merged head as local. (see another alternative I tried and failed https://github.com/teambit/bit/pull/5616).