fix: avoid showing errors when importing a component while its dependencies are modified (#5682)
The following error was showing up during `bit import` when a dependency was in the workspace in modified status:
```
unable to import the following components due to local changes, use --merge flag to merge your local changes or --override to override them
<comp-name>
```
This PR fixes it by skipping the dependencies-status check. It was relevant on legacy only.