improvement(bit-remove): deprecate "--delete-files" flag, it is the default now (#5747)
Users can opt-out by providing `--keep-files`.

The reason behind this change is that in almost all the cases users want to delete the component files as well.
Back then in the legacy it was different. Files normally were existing in the project and were added by `bit add` (vs. `bit create in Harmony).