fix: populate lanes data on ModelComponent object when it was rewritten (#5802)
The bug this is fixes is running `bit checkout latest --all` when on a lane and a component has no head. 
It shows the error `has no versions and the head is empty`. the reason is that during the import process, the model-component is rewritten to the filesystem and re-entered to the cache. Because it's in the cache, Bit assumes that the lanes data is already populated.