fix: ignore invalid bit object paths (#5797)
Currently, when a component object got a different path for some reason, it is still showing in `bit cat-scope`, it is indexed into the `index.json` file. However, when Bit tries to load it, it fails since the hash path doesn't exist.
This PR makes sure to never consider objects with malformed paths.