feat: enable serialization and deserialization of schema in one place (#5831)
Currently, each schema class needs to implement `toObject` and `fromObject` so then it's possible to pass the data via GraphQL and then deserialize it back to classes once received. 
It's a lot of work and is error-prone. Instead, using the class-transformer package, each class was configured with the data needs for the serialization, and is all taken care of automatically.