fix: avoid traversing tags on Harmony (#5863)
Currently, when traversing the component versions, once the traversal is done, it goes to each one of the tags and search for it as well. It was done originally to support legacy that didn't have the concept of head/parents.
This PR avoid this extra work if the component is Harmony.
In case when the traversal starts at a certain point, it also avoid errors of VersionNotFound when they're not part of the graph. (the attached e2e-test demonstrates it).