fix: consider locally-configured-remotes when resolving bit-id from scope (#5941)
currently, if the remote was added locally (via `bit remote add some-remote`) and the component doesn't exist in the local scope, it assumes the id doesn't have scope-name. 
with this fix, it'll check whether the first part of the id is the scope-name by comparing it to the local remotes.