fix: fix resolveComponentId when id has a non-default-scope and .bitmap has same name (#5977)
For example: 
The workspace has a component: `comp-a` with a default-scope `my-scope`, and we want to resolve a string-id: `another-scope/comp-a`.
Currently, `workspace.resolveComponentId` will return the comp-a from the .bitmap unexpectedly. 
This fix makes sure to not ignore the scope-name in this case because the defaultScope is not the same as the given scope.