fix(lanes): remove lane-key from .bitmap file when switching to main (#6012)
This fixes a bug where switching from a lane to main, it shows a success message but the lane was still checked out.
It happened when the lane-switch didn't make any change to the .bitmap other than the lane-key and lane-key removal didn't trigger a bitmap-change. As a result, the .bitmap lane-key was still pointing to the previous lane.