fix(bit-sign): avoid fetching dists of build-failed (#6027)
Currently if the component's build-status is failed, `bit sign` throws an error when on a lane:
```
ENOENT: no such file or directory, open '/tmp/scope-fs/objects/ed/af2594be2b2603149910c7ae86197e7c3d3975
```
Because the dists paths are saved into the model, Bit tries to fetch them when writing the capsule from the original-scope. 
This PR fixes it to not fetch the dists from the remote in case the build was not success. 
(if the build is success, then bit-sign won't try to get the dists anyway, so no issues there).