improvement: show the aspect+task that threw an error during build pipline (#6134)
currently, only when the task fails it shows the task details. If an error was thrown, it only show the error message, but it's hard to tell what task threw the error.
This PR catches the error coming from the task and shows the details as part of the build-pipeline output.