improvement: introduce helpUrl prop for commands (#6171)
A new prop `helpUrl` was added to the `Command` class. The CLI aspect is the one that responsible to concatenate the base-url from the community aspect to the relative-url provides by the `helpUrl` of the command. 
Since the output provided by `yargs` is already parsed and manipulated heavily, it was easy to let it render the URL as well (and with an underline style).