fix(bit-lane-diff): allow using "main" explicitly (#6224)
currently, an error is thrown when running `bit lane diff main`. This PR makes it the same as running `bit lane diff`.

Also, fix the order of the conflict messages. 
The current lane is always the `+` and the other lane is the `-`. 
If the syntax is `[from] [to]`, then the `from` is the `-` and the to is the `+`. 

Previously, it wasn't consistent, and in many cases showed the other way around.