fix(merge): avoid running package-installation for each one of the components individually, also, support --skip-dependency-installation flag (#6288)
This fix is for both `bit merge` and `bit lane merge`. Currently, it was looping throw the components, and running the package-manager on each one of them.
This PR fixes it to run it once and only if `--skip-dependency-installation` was not entered and only if no conflicts were found. 
Also, fixed an issue with pattern containing negation.