improvement(merge): support merging config of the unmerged head (#6296)
When merging lanes and conflicts found, the checked out versions of the components remain the same. as a result, when loading the components, it loads the config from the current version and disregard the config of the target version. 
This PR takes into account the target version when calculating the config.

It also fixes an issue when merging a lane and some components are not part of the workspace to merge them correctly.
(previously, it didn't try to do any diff/merge and always overrode with the target components).