fix(bit-reset): revert the config in .bitmap (#6358)
Currently, if a config was added to the .bitmap file, it gets deleted during `bit tag`. as a result, when running `bit reset`, there is no way to get it back, and the workspace is left without the config changes unexpectedly. 
This PR implements a new mechanism of storing this config between the tag and the export, so then, `bit reset` can copy the config from this file and put it back in the .bitmap file.
See https://github.com/teambit/bit/discussions/5074#discussioncomment-1787197 for more info about this topic.