feat(bit-lane-merge): support merging unrelated components (#6380)
This is happening when a component was created on a lane and later another component with the same name and scope was exported to the remote. As a result, the lane-scope and the component-scope hold different components with the same name+scope.
This PR introduces a new flag `--resolve-unrelated` to help merging such a component.
The component history is left the same as in the origin (main), and once snapped, the `Version` object gets a new prop "unrelated", which holds a reference to the component on the lane. 
The content of the component is left as is by default, unless the user entered `theirs` as the merge-strategy.