fix(dependencies-resolution): consider variants/bitmap-config policy (#6385)
This is a regression caused by https://github.com/teambit/bit/pull/6281.
When a component-dependency is not part of the workspace and it's not in the workspace.jsonc, then, the version is resolved from the model. 
This PR fixes it to take into account "overrides", which are the bitmap-config / component.json / workspace.jsonc-overrides.
If the component is specified there, it uses the package.json in the node_modules to resolve the version from.