improvement(bit-checkout-head): sync non-exported components with their remote default-scope (#6396)
In some cases, the .bitmap is not up to date with exported component. Locally, a component is new, but in fact the component has been exported to the remote.
This could happen for example when the CI tags+exports and for some reason the .bitmap wasn't pushed to Git. 
With this PR, these new components are updated during `bit checkout head` to the latest version on the remote.