fix(bit-lane-merge): avoid damaging the workspace if the auto-snap fails (#6486)
currently, if during the lane-merge the auto-snap failed, it leaves the workspace is a state where the objects are written but the .bitmap is not. This PR writes the bitmap changes before the auto-snap along with the objects and lane save. 
Also it catches the error coming from the auto-snap and shows them in the final output.