refactor: remove onTag slot, call builder directory during tag/snap (#6489)
Previously, the tag was on legacy and to get the builder aspect running during tag, a few indirections was needed. There indirections weren't trivial and confusing. (tag called legacy-scope, which had the harmony-scope register to its onTag. this onTag registered the builder tagListener which eventually got called during tag).
Now it's easier to just call the builder directly from the Snapping aspect.