fix: avoid throwing ComponentNotFound when .bitmap has a non-exist version on the scope (#6496)
Happens when an exported component has a local tag and `.bit` got deleted. The local scope doesn't have this local tag anymore and it never reached the remote. 
Previously, this use-case was handled by the out-of-sync algorithm, which synced the version according to the remote.
Since Harmony, a new load-hook was added before the out-of-sync logic, which imports components from remotes. This import threw ComponentNotFound in this case. 
This PR fixes it by running some of the out-of-sync before the on-load hook gets called.